pipelines/microsoft_gaming/requirements.txt
pandas>=1.5.0
numpy>=1.24.0
requests>=2.28.0
PyYAML>=6.0
scipy>=1.10.0
pytest>=7.0.0
pytest-cov>=4.0.0
